move trial data to story
